+++ source/Modules/GuestConfigurationPolicy/Public/New-GuestConfigurationDeployPolicyDefinition.ps1
@@ -35,6 +35,7 @@ function New-GuestConfigurationDeployPolicyDefinition
         [String]
         $ContentHash,
 
+        [Parameter()]
         [AssignmentType]
         $AssignmentType,
 
@@ -222,7 +223,7 @@ function New-GuestConfigurationDeployPolicyDefinition
                     version         = $ConfigurationVersion
                     contentUri      = "[parameters('contentUri')]"
                     contentHash     = "[parameters('contentHash')]"
-                    assignmentType  = "[parameters('assignmentType')]"
+                    assignmentType  = "$AssignmentType"
                     configurationParameter = $ParameterMapping
                 }
             }
@@ -239,7 +240,7 @@ function New-GuestConfigurationDeployPolicyDefinition
                     name        = "[parameters('configurationName')]"
                     contentUri  = "[parameters('contentUri')]"
                     contentHash = "[parameters('contentHash')]"
-                    assignmentType  = "[parameters('assignmentType')]"
+                    assignmentType  = "$AssignmentType"
                     version     = $ConfigurationVersion
                     configurationParameter = $ParameterMapping
                 }