+++ source/Public/Install-GuestConfigurationAgent.ps1
@@ -14,8 +14,16 @@ function Install-GuestConfigurationAgent
     $gcBinPath = Get-GuestConfigBinaryPath
     $gcBinRootPath = Get-GuestConfigBinaryRootPath
     $OsPlatform = Get-OSPlatform
+    if ($PSBoundParameters.ContainsKey('Force') -and $PSBoundParameters['Force'])
+    {
+        $withForce = $true
+    }
+    else
+    {
+        $withForce = $false
+    }
 
-    if ((-not (Test-Path -Path $gcBinPath)) -or $Force.IsPresent)
+    if ((-not (Test-Path -Path $gcBinPath)) -or $withForce)
     {
         # Clean the bin folder
         Write-Verbose -Message "Removing existing installation from '$gcBinRootPath'."
@@ -49,8 +57,6 @@ function Install-GuestConfigurationAgent
                 chmod @('+x', $_.FullName)
             }
         }
-
-        Install-GuestConfigurationMonkeyPatch
     }
     else
     {